fresh_traffic_multimodal_research/requirements.txt
pyyaml>=6.0
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
matplotlib>=3.9
transformers>=4.48
datasets>=3.0
torch>=2.3
opencv-python>=4.10
ultralytics>=8.3
tqdm>=4.66
